requirements.txt
    fastapi==0.111.0
    uvicorn==0.30.1
    sqlalchemy==2.0.30
    psycopg2-binary==2.9.9
    python-jose[cryptography]==3.3.0
    passlib[bcrypt]==1.7.4
    python-multipart==0.0.9
    psutil==5.9.8
    prometheus_client==0.20.0